% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decentlab_variables_look_up.R
\name{decentlab_variables_look_up}
\alias{decentlab_variables_look_up}
\title{Function to provide a look-up table of variable names for switching variable
names and potential variable filtering for Decentlab API data tables.}
\usage{
decentlab_variables_look_up()
}
\value{
Tibble.
}
\description{
Function to provide a look-up table of variable names for switching variable
names and potential variable filtering for Decentlab API data tables.
}
\author{
Stuart K. Grange
}
