% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_decentlab_time_series.R
\name{export_decentlab_time_series}
\alias{export_decentlab_time_series}
\title{Function to get data from the Decentlab API and export daily files to a 
directory.}
\usage{
export_decentlab_time_series(
  domain,
  key,
  device,
  start,
  end,
  directory,
  progress = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{domain}{A Decentlab API domain.}

\item{key}{A Decentlab API key for \code{domain}.}

\item{device}{A vector of device codes.}

\item{start}{Start date to get time series for.}

\item{end}{End date to get time series for}

\item{directory}{Directory where the files should be exported to.}

\item{progress}{Should the function show a progress bar?}

\item{verbose}{Should the functions give messages?}
}
\value{
Invisible \code{domain}.
}
\description{
Function to get data from the Decentlab API and export daily files to a 
directory.
}
\seealso{
\code{\link{get_decentlab_time_series}}
}
\author{
Stuart K. Grange
}
