% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decent_lab_last_values.R
\name{get_decent_lab_last_values}
\alias{get_decent_lab_last_values}
\title{Function to get observational data from the Decentlab API.}
\usage{
get_decent_lab_last_values(domain, key)
}
\arguments{
\item{domain}{A Decentlab API domain.}

\item{key}{A Decentlab API key for \code{domain}.}
}
\value{
Tibble.
}
\description{
Function to get observational data from the Decentlab API.
}
\author{
Stuart K. Grange
}
