% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decent_lab_time_series.R
\name{get_decent_lab_time_series}
\alias{get_decent_lab_time_series}
\title{Function to get observational data from the Decentlab API.}
\usage{
get_decent_lab_time_series(
  domain,
  key,
  device,
  sensor = NA,
  start = NA,
  as_wide = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{domain}{A Decentlab API domain.}

\item{key}{A Decentlab API key for \code{domain}.}

\item{device}{A vector of device codes.}

\item{sensor}{A vector of device's sensor codes.}

\item{start}{Start date to get time series for.}

\item{as_wide}{Should the return be in "wide" format?}

\item{verbose}{Should the functions give messages?}
}
\value{
Tibble.
}
\description{
Function to get observational data from the Decentlab API.
}
\seealso{
\code{\link{get_decent_lab_last_values}}
}
\author{
Stuart K. Grange
}
