% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decentlab_last_values.R
\name{get_decentlab_last_values}
\alias{get_decentlab_last_values}
\title{Function to get observational data from the Decentlab API.}
\usage{
get_decentlab_last_values(domain, key)
}
\arguments{
\item{domain}{A Decentlab API domain.}

\item{key}{A Decentlab API key for \code{domain}.}
}
\value{
Tibble.
}
\description{
Function to get observational data from the Decentlab API.
}
\examples{

# Get last values for Dencentlab's demo user
get_decentlab_last_values(
  domain = "demo.decentlab.com",
  key = "eyJrIjoiclhMRFFvUXFzQXpKVkZydm52b0VMRVg3M3U2b3VqQUciLCJuIjoiZGF0YS1xdWVyeS1hcGktZGVtby0yIiwiaWQiOjF9"
)

}
\author{
Stuart K. Grange
}
