% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_decentlab_api_files.R
\name{list_decentlab_api_files}
\alias{list_decentlab_api_files}
\title{Function to list and get information from a collection of exported Decentlab 
API files.}
\usage{
list_decentlab_api_files(path, select = TRUE)
}
\arguments{
\item{path}{Path/directory where exported Decentlab API files are located.}

\item{select}{Should only a selected set of variables be returned?}
}
\value{
Tibble.
}
\description{
\code{list_decentlab_api_files} will extract information from file names to
allow for easy filtering of the file list.
}
\author{
Stuart K. Grange
}
