% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_decentlab_api_export.R
\name{read_decentlab_api_export}
\alias{read_decentlab_api_export}
\title{Function to read exported Decentlab API files with options to add site and 
sensing element IDs.}
\usage{
read_decentlab_api_export(
  file,
  df_site_ranges = NA,
  df_sensing_elements_ranges = NA,
  variable_switch = FALSE,
  date_round = FALSE,
  calculate_date = FALSE,
  warn = TRUE,
  progress = FALSE
)
}
\arguments{
\item{file}{Vector of file names of files written by 
\code{\link{export_decentlab_time_series}}.}

\item{df_site_ranges}{An optional data frame containing site information.}

\item{df_sensing_elements_ranges}{An optional data frame containing sensing
element information.}

\item{variable_switch}{Should the variables be switched to "clean" names and
filtered to those contained in \code{\link{read_decentlab_variables}}.}

\item{date_round}{Should the dates be rounded to seconds? Sometimes, the API
returns data with sub-second accuracy.}

\item{calculate_date}{Should the \code{date} variable calculated from the 
API's date which represents the date when the data were inserted into the
remote database and therefore represent the end of the measurement period 
with a minor delay? The date from the API is renamed to \code{date_end} 
(after being floor rounded) and both \code{date_round} and 
\code{calculate_date} cannot be \code{TRUE}.}

\item{warn}{Should the function raise warnings?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Tibble.
}
\description{
Function to read exported Decentlab API files with options to add site and 
sensing element IDs.
}
\seealso{
\code{\link{export_decentlab_time_series}}, 
\code{\link{read_decentlab_variables}}.
}
\author{
Stuart K. Grange.
}
