% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_decentlab_variables.R
\name{read_decentlab_variables}
\alias{read_decentlab_variables}
\title{Function to read a provided look-up table of variable names for the 
Decentlab API.}
\usage{
read_decentlab_variables()
}
\value{
Tibble.
}
\description{
\code{read_decentlab_variables}'s return can be used for for switching 
variable names and variable filtering for Decentlab API data tables.
}
\author{
Stuart K. Grange
}
