% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_decentlab_time_series.R
\name{export_decentlab_time_series}
\alias{export_decentlab_time_series}
\title{Function to get data from the Decentlab API and export daily files to a 
directory.}
\usage{
export_decentlab_time_series(
  domain,
  key,
  device,
  start,
  end,
  as_wide = TRUE,
  directory,
  warn = TRUE,
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{domain}{A Decentlab API domain.}

\item{key}{A Decentlab API key for \code{domain}.}

\item{device}{A vector of device codes.}

\item{start}{Start date to get time series for.}

\item{end}{End date to get time series for.}

\item{as_wide}{Should the return be in "wide" format? The \code{as_wide} 
argument will not be honoured if the device's data has a \code{channel} 
variable because the wide format does not work well when this additional 
identifier is present.}

\item{directory}{Directory/path where the files should be exported to.}

\item{warn}{Should the function raise warnings?}

\item{verbose}{Should the functions give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Invisible \code{domain}.
}
\description{
Function to get data from the Decentlab API and export daily files to a 
directory.
}
\seealso{
\code{\link{get_decentlab_time_series}}
}
\author{
Stuart K. Grange
}
