% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decent_lab_time_series.R
\name{get_decent_lab_time_series}
\alias{get_decent_lab_time_series}
\title{Function to get observational data from the Decentlab API.}
\usage{
get_decent_lab_time_series(
  domain,
  key,
  device,
  sensor = NA,
  start = NA,
  end = NA,
  as_wide = FALSE,
  tz = "UTC",
  verbose = FALSE
)
}
\arguments{
\item{domain}{A Decentlab API domain.}

\item{key}{A Decentlab API key for \code{domain}.}

\item{device}{A vector of device codes.}

\item{sensor}{A vector of device's sensor codes.}

\item{start}{Start date to get time series for.}

\item{end}{End date to get time series for}

\item{as_wide}{Should the return be in "wide" format?}

\item{tz}{Time-zone for the time series to be returned in.}

\item{verbose}{Should the functions give messages?}
}
\value{
Tibble.
}
\description{
Function to get observational data from the Decentlab API.
}
\examples{

# Get a device's observations for a short time period
get_decent_lab_time_series(
  domain = "demo.decentlab.com",
  key = "eyJrIjoiclhMRFFvUXFzQXpKVkZydm52b0VMRVg3M3U2b3VqQUciLCJuIjoiZGF0YS1xdWVyeS1hcGktZGVtby0yIiwiaWQiOjF9",
  device = 3001,
  start = lubridate::today() - lubridate::days(1),
  end = lubridate::today() + lubridate::days(1)
)

# Just get CO2 for the sensor
get_decent_lab_time_series(
  domain = "demo.decentlab.com",
  key = "eyJrIjoiclhMRFFvUXFzQXpKVkZydm52b0VMRVg3M3U2b3VqQUciLCJuIjoiZGF0YS1xdWVyeS1hcGktZGVtby0yIiwiaWQiOjF9",
  device = 3001,
  sensor = "senseair-lp8-co2",
  start = lubridate::today() - lubridate::days(1),
  end = lubridate::today() + lubridate::days(1)
)

# Get a device's observations for a short time period in wide format
get_decent_lab_time_series(
  domain = "demo.decentlab.com",
  key = "eyJrIjoiclhMRFFvUXFzQXpKVkZydm52b0VMRVg3M3U2b3VqQUciLCJuIjoiZGF0YS1xdWVyeS1hcGktZGVtby0yIiwiaWQiOjF9",
  device = 3001,
  start = lubridate::today() - lubridate::days(1),
  end = lubridate::today() + lubridate::days(1),
  as_wide = TRUE
)

}
\seealso{
\code{\link{get_decent_lab_last_values}}
}
\author{
Stuart K. Grange
}
