% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_decentlab_api_export.R
\name{read_decentlab_api_export}
\alias{read_decentlab_api_export}
\title{Function to read exported Decentlab API files with options to add site and 
sensing element IDs.}
\usage{
read_decentlab_api_export(
  file,
  df_site_ranges = NA,
  df_sensing_elements_ranges = NA,
  variable_switch = FALSE,
  date_round = FALSE,
  progress = FALSE
)
}
\arguments{
\item{file}{Vector of file names of files written by 
\code{\link{export_decentlab_time_series}}.}

\item{df_site_ranges}{An optional data frame containing site information.}

\item{df_sensing_elements_ranges}{An optional data frame containing sensing
element information.}

\item{variable_switch}{Should the variables be switched to "clean" names and
filtered to those contained in \code{\link{decentlab_variables_look_up}}.}

\item{date_round}{Should the dates be rounded to seconds? Sometimes, the API
returns data with sub-second accuracy.}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Tibble.
}
\description{
Function to read exported Decentlab API files with options to add site and 
sensing element IDs.
}
\seealso{
\code{\link{export_decentlab_time_series}}
}
\author{
Stuart K. Grange.
}
